moved spinner css from app.css to components/Spinner.css
